Add frexp implementation and tests

+++ userland/tests/libm/common/mtest.h
@@ -79,7 +79,7 @@ struct ff_f {POS int r; unsigned int x; unsigned int x2; unsigned int y; unsigne
 struct dd_d {POS int r; double x; double x2; double y; float dy; int e; };
 struct ll_l {POS int r; long double x; long double x2; long double y; float dy; int e; };
 struct d_di {POS int r; double x; double y; float dy; long long i; int e; };
-struct f_fi {POS int r; float x; float y; float dy; long long i; int e; };
+struct f_fi {POS int r; unsigned int x; unsigned int y; unsigned int dy; long long i; int e; };
 struct l_li {POS int r; long double x; long double y; float dy; long long i; int e; };
 struct di_d {POS int r; double x; long long i; double y; float dy; int e; };
 struct fi_f {POS int r; unsigned int x; long long i; unsigned int y; unsigned int dy; int e; };
